backend/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
pylusat==0.5.8
geopandas==0.14.3
pyproj==3.7.2
rasterio==1.5.0
numpy==2.4.4
shapely==2.1.2
pyshp==3.0.3
psycopg[binary]
SQLAlchemy==2.0.29
GeoAlchemy2==0.14.7
python-dotenv==1.0.1
